adds collection
