--- src/Query.rs
@@ -1,32 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug)]
-struct Url {
-    raw: String,
-    protocol: String,
-    host: Vec<String>,
-    path: Vec<String>,
-}
-#[derive(Serialize, Deserialize, Debug)]
-struct Graphql {
-    query: String,
-    variables: String,
-}
-#[derive(Serialize, Deserialize, Debug)]
-struct Body {
-    mode: String,
-    graphql: Graphql,
-}
-#[derive(Serialize, Deserialize, Debug)]
-struct Request {
-    method: String,
-    header: Vec<String>,
-    body: Body,
-    url: Url,
-}
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Query {
-    name: String,
-    request: Request,
-    response: Vec<String>,
-}
