+++ src/lib.rs
@@ -1,3 +1,3 @@
-//pub mod Query;
 pub mod parser;
+pub mod collection;
 //pub mod writer;
