+++ tests/parser.rs
@@ -5,3 +5,4 @@ fn valid_collection_json() {
     // todo: break up parser into parseing and file creation
     // test parser output against a mocked valid postman.collection.json
 }
+
